Change PointSize from half to float

+++ Shaders/Disk.cginc
@@ -6,7 +6,7 @@
 
 // Uniforms
 half4 _Tint;
-half _PointSize;
+float _PointSize;
 float4x4 _Transform;
 
 #if _COMPUTE_BUFFER
