+++ Shaders/Point4D.shader
@@ -41,7 +41,7 @@ Shader "Point Cloud/Point4D"
             {
                 float4 position : SV_Position;
                 half3 color : COLOR;
-                half psize : PSIZE;
+                float psize : PSIZE;
                 UNITY_FOG_COORDS(0)
             };
 
@@ -53,7 +53,7 @@ Shader "Point Cloud/Point4D"
 
             half4 _Tint;
             float4x4 _Transform;
-            half _PointSize;
+            float _PointSize;
             float _Chiral;
 
             float4x4 _Rotation4D;
